% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineWeightsAndNorms.R
\name{combineWeightsAndNorms}
\alias{combineWeightsAndNorms}
\title{combineWeightsAndNorms}
\usage{
combineWeightsAndNorms(weights, norms)
}
\arguments{
\item{weights}{a list of weights generated by \code{\link{getFAWeights}}}

\item{norms}{a data.frame of normalised indicator values generated by \code{\link{normalScores}}}
}
\description{
This function takes the normalised indicator scores and the weights derived from
factor analysis, multiplies them out and then takes the sum of these weighted
indicator scores to get the final score for that domain.
The function takes a list of weights (generated by \code{\link{getFAWeights}}) and a
data.frame of normalised scores, it returns a numeric vector containing the
combined domain score.
}
\keyword{SIMD,}
\keyword{openSIMD,}
\keyword{simdr}
