% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expoTransform.R
\name{expoTransform}
\alias{expoTransform}
\title{combineWeightsAndNorms}
\usage{
expoTransform(ranks)
}
\arguments{
\item{ranks}{a numeric or integer vector of domain ranks.}
}
\description{
This function recapitulates the exponential transformation used in SIMD to
transform the weighted domain ranks before combining domains.
The function takes an numeric vector (of the ranks) and returns a numeric vector
of equal length containing the transformed values.
}
\keyword{SIMD,}
\keyword{openSIMD,}
\keyword{simdr}
