% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFAWeights.R
\name{getFAWeights}
\alias{getFAWeights}
\title{replaceMissing}
\usage{
getFAWeights(dat, ...)
}
\arguments{
\item{dat}{a data frame containing the normalised indicator variables that need
to be weighted, and nothing else.}

\item{...}{Other arguments to be passed to \code{\link[psych]{fa}}}
}
\description{
This function performs factor analysis and returns the proportional scores or
weights for each variable on the first factor. This becomes the weight for
combining the normalised indicator scores.
The function takes a data frame, assumes that all variables are being processed
in factor analysis and returns a list where each element is a numeric value,
the weight corresponding to that column in the data. Factor analysis is performed
by \code{\link[psych]{fa}}
}
\keyword{SIMD,}
\keyword{openSIMD,}
\keyword{simdr}
