% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_qualifications.R
\docType{data}
\name{no_qualifications}
\alias{no_qualifications}
\title{Working age adults with no qualifications: standardised ratio}
\format{A data frame with 6976 rows and 3 variables:
\describe{
  \item{DZ}{the data zone}
  \item{Noquals}{Count of individuals with no qualifications}
  \item{nnoquals}{Normalised count, derived from SIMD proceedure in SAS}
  ...
}}
\source{
\url{http://www.gov.scot/Topics/Statistics/SIMD}
}
\usage{
no_qualifications
}
\description{
This is one of the indicators in SIMD and is included for the purposes of testing our normalisation proceedure.
}
\details{
Please see the technical notes \url{http://www.gov.scot/Resource/0050/00504822.pdf} for full information on all of the data in SIMD.
}
\keyword{datasets}
