% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceMissing.R
\name{replaceMissing}
\alias{replaceMissing}
\title{replaceMissing}
\usage{
replaceMissing(v)
}
\arguments{
\item{v}{a numeric vector as the input variable}
}
\description{
Replace missing values in normalised indicators.
This is a simple wrapper around replace and should be called after \code{normalScores}.
Note that \code{normalScores} will show a warning when NA values are present in the data.
This function is designed to assign the missing values to the centre of the normalised distribution (0).
}
\examples{
replaceMissing(c(1, 2, 3, NA))

}
\keyword{SIMD,}
\keyword{openSIMD,}
\keyword{simdr}
