% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas_simd_domains.R
\docType{data}
\name{sas_simd_domains}
\alias{sas_simd_domains}
\title{Domain rankings for SIMD 16 derived from SAS proceedure}
\format{A data frame with 6976 rows and 13 variables:
\describe{
  \item{DZ}{Data Zone}
  \item{IZ}{}
  \item{LA}{Local Authority}
  \item{pop}{Population}
  \item{wapop}{Working age Population}
  \item{incrnk}{Income Domain Rank}
  \item{emprnk}{Employment Domain Rank}
  \item{hearnk}{Health Domain Rank}
  \item{edurnk}{Education Domain Rank}
  \item{accrnk}{Access Domain Rank}
  \item{crirnk}{Crime Domain Rank}
  \item{hournk}{Housing Domain Rank}
  \item{SIMD}{Overall SIMD Rank}
  ...
}}
\source{
\url{http://www.gov.scot/Topics/Statistics/SIMD}
}
\usage{
sas_simd_domains
}
\description{
The published ranks of the individual domains and the overall SIMD rank for 2016 calculated in SAS.
We include this data for the purposes of measuring the correspondence of ranks calculated in R with ranks calculated in SAS.
}
\details{
Please see the technical notes \url{http://www.gov.scot/Resource/0050/00504822.pdf} for full information on all of the data in SIMD.
}
\keyword{datasets}
