% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas_weights.R
\docType{data}
\name{sas_weights}
\alias{sas_weights}
\title{Factor analysis weights derived from SAS}
\format{A data frame with ___ rows and ___ variables:
\describe{
  \item{variable}{description}
  \item{}{Count of individuals with no qualifications}
  \item{}{Normalised count, derived from SIMD proceedure in SAS}
  ...
}}
\source{
\url{http://www.gov.scot/Topics/Statistics/SIMD}
}
\usage{
sas_weights
}
\description{
[describe it here]
}
\details{
Please see the technical notes \url{http://www.gov.scot/Resource/0050/00504822.pdf} for full information on all of the data in SIMD.
}
\keyword{datasets}
