% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simd16_indicators.R
\docType{data}
\name{simd16_indicators}
\alias{simd16_indicators}
\title{Indicator data from published SIMD 2016}
\format{A data frame with 6976 rows and 36 variables:
\describe{
  \item{Data_Zone}{description}
  \item{Intermediate_Zone}{description}
  \item{Council_area}{description}
  \item{Total_population}{description}
  \item{Working_age_population_revised}{description}
  \item{Income_rate}{description}
  \item{Income_count}{description}
  \item{Employment_rate}{description}
  \item{Employment_count}{description}
  \item{CIF}{description}
  \item{ALCOHOL}{description}
  \item{DRUG}{description}
  \item{SMR}{description}
  \item{DEPRESS}{description}
  \item{LBWT}{description}
  \item{EMERG}{description}
  \item{Attendance}{description}
  \item{Attainment}{description}
  \item{Noquals}{description}
  \item{NEET}{description}
  \item{HESA}{description}
  \item{drive_petrol}{description}
  \item{drive_GP}{description}
  \item{drive_PO}{description}
  \item{drive_primary}{description}
  \item{drive_retail}{description}
  \item{drive_secondary}{description}
  \item{PT_GP}{description}
  \item{PT_Post}{description}
  \item{PT_retail}{description}
  \item{crime_count}{description}
  \item{crime_rate}{description}
  \item{overcrowded_count}{description}
  \item{nocentralheat_count}{description}
  \item{overcrowded_rate}{description}
  \item{nocentralheat_rate}{description}
  ...
}}
\source{
\url{http://www.gov.scot/Topics/Statistics/SIMD}
}
\usage{
simd16_indicators
}
\description{
We include the indicator data from the 2016 SIMD for purposes of evaluating the correspondence between the SAS and R methodologies.
To see our quality control analysis see [quality assurance section of the openSIMD documentation](https://thedatalabscotland.github.io/openSIMD_site/QA.html)
}
\details{
Please see the technical notes \url{http://www.gov.scot/Resource/0050/00504822.pdf} for full information on all of the data in SIMD.
}
\keyword{datasets}
