% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simd16_indicators.R
\docType{data}
\name{simd16_indicators}
\alias{simd16_indicators}
\title{Indicator data from published SIMD 2016}
\format{A data frame with ___ rows and ___ variables:
\describe{
  \item{variable}{description}
  \item{}{Count of individuals with no qualifications}
  \item{}{Normalised count, derived from SIMD proceedure in SAS}
  ...
}}
\source{
\url{http://www.gov.scot/Topics/Statistics/SIMD}
}
\usage{
simd16_indicators
}
\description{
[describe it here]
}
\details{
Please see the technical notes \url{http://www.gov.scot/Resource/0050/00504822.pdf} for full information on all of the data in SIMD.
}
\keyword{datasets}
